// src/components/TimeLeft/time-left.types.tsx
export interface TimeLeftProps {
    targetDate: string
}

export interface ColumnProps {
    name: string;
    number?: string;
}

export interface CountdownSectionProps {
    months: string;
    days: string;
    hours: string;
    minutes: string;
    seconds: string;
    weeks: string;
    years: string;
}